third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/cross-origin-unsupported-url-expected.txt

CONSOLE WARNING: line 14: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
CONSOLE ERROR: line 20: Failed to load ftp://127.0.0.1/: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: line 1: Failed to load ftp://127.0.0.1/: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: line 20: Failed to load ftp://127.0.0.1/: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: line 1: Failed to load ftp://127.0.0.1/: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: line 20: Failed to load localhost:8080/: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: line 1: Failed to load localhost:8080/: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: line 20: Failed to load localhost:8080/: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: line 1: Failed to load localhost:8080/: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: line 20: Failed to load tel:1234: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: line 1: Failed to load tel:1234: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: line 20: Failed to load tel:1234: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: line 1: Failed to load tel:1234: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
Test cross-origin XHRs to CORS-unsupported protocol schemes in the URL.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS errorEvent.name is "NetworkError"
PASS errorEvent.type is "error"
PASS xhr.send() did not throw exception.
PASS errorEvent.name is "NetworkError"
PASS errorEvent.type is "error"
PASS xhr.send() did not throw exception.
PASS errorEvent.name is "NetworkError"
PASS errorEvent.type is "error"
PASS xhr.send() did not throw exception.
PASS errorEvent.name is "NetworkError"
PASS errorEvent.type is "error"
PASS xhr.send() did not throw exception.
PASS errorEvent.name is "NetworkError"
PASS errorEvent.type is "error"
PASS xhr.send() did not throw exception.
PASS errorEvent.name is "NetworkError"
PASS errorEvent.type is "error"
PASS xhr.send() did not throw exception.
PASS successfullyParsed is true

TEST COMPLETE
